add ActionDSL marker
