+++ src/main/kotlin/ca/warp7/frc/action/ActionBuilder.kt
@@ -1,9 +1,11 @@
 package ca.warp7.frc.action
 
+@ActionDSL
 interface ActionBuilder {
 
     /**
      * Adds an action to the manager scope
      */
+    @ActionDSL
     operator fun Action.unaryPlus()
 }
