+++ src/main/kotlin/ca/warp7/frc/action/ActionDSL.kt
@@ -1,0 +1,4 @@
+package ca.warp7.frc.action
+
+@DslMarker
+annotation class ActionDSL
