+++ src/main/kotlin/ca/warp7/frc/action/ActionLib.kt
@@ -6,21 +6,25 @@ package ca.warp7.frc.action
 /**
  * Returns an action that runs other actions sequentially
  */
+@ActionDSL
 fun sequential(block: ActionBuilder.() -> Unit): Action = Sequential().apply(block)
 
 /**
  * Returns an action that runs other actions in parallel
  */
+@ActionDSL
 fun parallel(block: ActionBuilder.() -> Unit): Action = Parallel().apply(block)
 
 /**
  * Returns an action that does nothing until a condition is met
  */
+@ActionDSL
 fun waitUntil(predicate: (elapsed: Double) -> Boolean): Action = WaitUntil(predicate)
 
 /**
  * Returns an action that runs only once
  */
+@ActionDSL
 inline fun runOnce(crossinline block: () -> Unit) = object : Action {
     override fun name() = "runOnce"
     override fun shouldFinish() = true
@@ -30,6 +34,7 @@ inline fun runOnce(crossinline block: () -> Unit) = object : Action {
 /**
  * Returns an action that gets called periodically forever
  */
+@ActionDSL
 inline fun periodic(crossinline block: () -> Unit) = object : Action {
     override fun name() = "periodic"
     override fun shouldFinish() = false
@@ -39,4 +44,5 @@ inline fun periodic(crossinline block: () -> Unit) = object : Action {
 /**
  * Returns an action that does nothing for [seconds] seconds
  */
+@ActionDSL
 fun wait(seconds: Number) = waitUntil { elapsed -> elapsed > seconds.toDouble() }
